+++ src/base64_format.rs
@@ -1,9 +1,15 @@
 #[cfg(feature = "serdejson")]
 use base64::{decode, encode, DecodeError};
 #[cfg(feature = "serdejson")]
+use paste;
+#[cfg(feature = "serdejson")]
+use regex::Regex;
+#[cfg(feature = "serdejson")]
 use serde::de::{Deserialize, Deserializer, Error};
 #[cfg(feature = "serdejson")]
 use serde::ser::{Serialize, Serializer};
+#[cfg(feature = "serdejson")]
+use serde_valid;
 use std::ops::{Deref, DerefMut};
 
 #[derive(Debug, Clone, PartialEq, PartialOrd)]
@@ -34,6 +40,51 @@ impl<'de> Deserialize<'de> for ByteArray {
     }
 }
 
+// Validation macro to create impls for serde_valid integration.
+macro_rules! impl_validate_byte_array {
+    (
+        $ErrorType:ident,
+        $DiveFn:expr,
+        $limit_type:ty$(,)*
+
+    ) => {
+        paste::paste! {
+            #[cfg(feature = "serdejson")]
+            impl serde_valid::[<Validate $ErrorType >] for ByteArray {
+                fn [<validate_ $ErrorType:snake>](&self, limit: $limit_type) -> Result<(), serde_valid::[<$ErrorType Error>]> {
+                    self.$DiveFn.[<validate_ $ErrorType:snake>](limit)
+                }
+            }
+        }
+    };
+}
+
+// Allow validation of encoded string.
+impl_validate_byte_array!(Pattern, to_string(), &Regex);
+impl_validate_byte_array!(MaxLength, to_string(), usize);
+impl_validate_byte_array!(MinLength, to_string(), usize);
+
+#[cfg(feature = "serdejson")]
+impl serde_valid::ValidateEnumerate<&'static str> for ByteArray {
+    fn validate_enumerate(
+        &self,
+        enumerate: &[&'static str],
+    ) -> Result<(), serde_valid::EnumerateError> {
+        self.to_string().validate_enumerate(enumerate)
+    }
+}
+
+// Also allow validation decoded internals.
+impl_validate_byte_array!(MaxItems, 0, usize);
+impl_validate_byte_array!(MinItems, 0, usize);
+
+#[cfg(feature = "serdejson")]
+impl serde_valid::ValidateUniqueItems for ByteArray {
+    fn validate_unique_items(&self) -> Result<(), serde_valid::UniqueItemsError> {
+        self.0.validate_unique_items()
+    }
+}
+
 impl std::str::FromStr for ByteArray {
     type Err = DecodeError;
 
@@ -58,5 +109,87 @@ impl Deref for ByteArray {
 impl DerefMut for ByteArray {
     fn deref_mut(&mut self) -> &mut Vec<u8> {
         &mut self.0
+    }
+}
+
+#[cfg(test)]
+#[cfg(feature = "serdejson")]
+mod serde_tests {
+    use super::*;
+    use serde_json::json;
+    use serde_valid::Validate;
+    use std::str::FromStr;
+
+    #[derive(Validate)]
+    struct ValidateByteArrayStruct {
+        // Validate encoded as string
+        #[validate(enumerate("YWJjZGU="))]
+        #[validate(max_length = 8)]
+        #[validate(min_length = 8)]
+        #[validate(pattern = ".*=")]
+        // Validate decoded as Vec (b"abcde")
+        #[validate(max_items = 5)]
+        #[validate(min_items = 5)]
+        #[validate(unique_items)]
+        byte_array: ByteArray,
+    }
+
+    #[test]
+    fn valid_byte_array() {
+        let test_struct = ValidateByteArrayStruct {
+            byte_array: ByteArray::from_str("YWJjZGU=").unwrap(),
+        };
+        assert!(test_struct.validate().is_ok());
+    }
+
+    #[test]
+    fn invalid_few_byte_array() {
+        let test_struct = ValidateByteArrayStruct {
+            byte_array: ByteArray::from_str("ZmZm").unwrap(),
+        };
+        let errors = test_struct.validate().unwrap_err().to_string();
+        assert_eq!(
+            errors,
+            json!({"errors":[],
+                "properties":{
+                    "byte_array":{
+                        "errors":[
+                            "The value must be in [YWJjZGU=].",
+                            "The length of the value must be `>= 8`.",
+                            "The value must match the pattern of \".*=\".",
+                            "The length of the items must be `>= 5`.",
+                            "The items must be unique."
+                            ]
+                        }
+                    }
+                }
+            )
+            .to_string()
+        );
+    }
+
+    #[test]
+    fn invalid_many_byte_array() {
+        let test_struct = ValidateByteArrayStruct {
+            byte_array: ByteArray::from_str("ZmZmZmZmZg==").unwrap(),
+        };
+        let errors = test_struct.validate().unwrap_err().to_string();
+        assert_eq!(
+            errors,
+            json!({"errors":[],
+                "properties":{
+                    "byte_array":{
+                        "errors":[
+                            "The value must be in [YWJjZGU=].",
+                            "The length of the value must be `<= 8`.",
+                            "The length of the items must be `<= 5`.",
+                            "The items must be unique."
+                            ]
+                        }
+                    }
+                }
+            )
+            .to_string()
+        );
     }
 }